+++ problem4j-spring-web/src/main/java/io/github/problem4j/spring/web/resolver/ConstraintViolationProblemResolver.java
@@ -22,6 +22,7 @@ import static io.github.problem4j.spring.web.parameter.ViolationSupport.VALIDATI
 import io.github.problem4j.core.Problem;
 import io.github.problem4j.core.ProblemContext;
 import io.github.problem4j.spring.web.ProblemFormat;
+import io.github.problem4j.spring.web.config.ProblemBeanPostProcessor;
 import io.github.problem4j.spring.web.parameter.Violation;
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.ConstraintViolationException;
@@ -51,7 +52,7 @@ public class ConstraintViolationProblemResolver extends AbstractProblemResolver 
    * @since 1.2.0
    */
   public ConstraintViolationProblemResolver() {
-    this(ProblemFormat.identity());
+    super(ConstraintViolationException.class);
   }
 
   /**
@@ -59,7 +60,11 @@ public class ConstraintViolationProblemResolver extends AbstractProblemResolver 
    *
    * @param problemFormat the problem format to use
    * @since 1.2.0
+   * @deprecated since 3.1.0 as {@link ProblemBeanPostProcessor} now assigns the {@link
+   *     ProblemFormat} after construction; use {@link #ConstraintViolationProblemResolver()}
    */
+  @SuppressWarnings("removal")
+  @Deprecated(since = "3.1.0", forRemoval = true)
   public ConstraintViolationProblemResolver(ProblemFormat problemFormat) {
     super(ConstraintViolationException.class, problemFormat);
   }