+++ problem4j-spring-web/src/main/java/io/github/problem4j/spring/web/resolver/ErrorResponseProblemResolver.java
@@ -20,6 +20,7 @@ import io.github.problem4j.core.Problem;
 import io.github.problem4j.core.ProblemBuilder;
 import io.github.problem4j.core.ProblemContext;
 import io.github.problem4j.spring.web.ProblemFormat;
+import io.github.problem4j.spring.web.config.ProblemBeanPostProcessor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.web.ErrorResponseException;
@@ -45,7 +46,7 @@ public class ErrorResponseProblemResolver extends AbstractProblemResolver {
    * @since 1.2.0
    */
   public ErrorResponseProblemResolver() {
-    this(ProblemFormat.identity());
+    super(ErrorResponseException.class);
   }
 
   /**
@@ -53,7 +54,11 @@ public class ErrorResponseProblemResolver extends AbstractProblemResolver {
    *
    * @param problemFormat the problem format to use
    * @since 1.2.0
+   * @deprecated since 3.1.0 as {@link ProblemBeanPostProcessor} now assigns the {@link
+   *     ProblemFormat} after construction; use {@link #ErrorResponseProblemResolver()}
    */
+  @SuppressWarnings("removal")
+  @Deprecated(since = "3.1.0", forRemoval = true)
   public ErrorResponseProblemResolver(ProblemFormat problemFormat) {
     super(ErrorResponseException.class, problemFormat);
   }