+++ problem4j-spring-web/src/main/java/io/github/problem4j/spring/web/resolver/ServerErrorProblemResolver.java
@@ -22,6 +22,7 @@ import static io.github.problem4j.spring.web.parameter.ViolationSupport.NAME_EXT
 import io.github.problem4j.core.Problem;
 import io.github.problem4j.core.ProblemContext;
 import io.github.problem4j.spring.web.ProblemFormat;
+import io.github.problem4j.spring.web.config.ProblemBeanPostProcessor;
 import org.jspecify.annotations.Nullable;
 import org.springframework.core.MethodParameter;
 import org.springframework.core.annotation.AnnotationUtils;
@@ -54,7 +55,7 @@ public class ServerErrorProblemResolver extends AbstractProblemResolver {
    * @since 1.2.0
    */
   public ServerErrorProblemResolver() {
-    this(ProblemFormat.identity());
+    super(ServerErrorException.class);
   }
 
   /**
@@ -62,7 +63,11 @@ public class ServerErrorProblemResolver extends AbstractProblemResolver {
    *
    * @param problemFormat the problem format to use
    * @since 1.2.0
+   * @deprecated since 3.1.0 as {@link ProblemBeanPostProcessor} now assigns the {@link
+   *     ProblemFormat} after construction; use {@link #ServerErrorProblemResolver()}
    */
+  @SuppressWarnings("removal")
+  @Deprecated(since = "3.1.0", forRemoval = true)
   public ServerErrorProblemResolver(ProblemFormat problemFormat) {
     super(ServerErrorException.class, problemFormat);
   }