+++ problem4j-spring-web/src/main/java/io/github/problem4j/spring/web/resolver/ServletRequestBindingProblemResolver.java
@@ -33,6 +33,7 @@ import io.github.problem4j.core.Problem;
 import io.github.problem4j.core.ProblemBuilder;
 import io.github.problem4j.core.ProblemContext;
 import io.github.problem4j.spring.web.ProblemFormat;
+import io.github.problem4j.spring.web.config.ProblemBeanPostProcessor;
 import java.util.Locale;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -88,7 +89,7 @@ public class ServletRequestBindingProblemResolver extends AbstractProblemResolve
    * @since 1.2.0
    */
   public ServletRequestBindingProblemResolver() {
-    this(ProblemFormat.identity());
+    super(ServletRequestBindingException.class);
   }
 
   /**
@@ -96,7 +97,11 @@ public class ServletRequestBindingProblemResolver extends AbstractProblemResolve
    *
    * @param problemFormat the problem format to use
    * @since 1.2.0
+   * @deprecated since 3.1.0 as {@link ProblemBeanPostProcessor} now assigns the {@link
+   *     ProblemFormat} after construction; use {@link #ServletRequestBindingProblemResolver()}
    */
+  @SuppressWarnings("removal")
+  @Deprecated(since = "3.1.0", forRemoval = true)
   public ServletRequestBindingProblemResolver(ProblemFormat problemFormat) {
     super(ServletRequestBindingException.class, problemFormat);
   }